% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet_connect.R
\name{wallet_connectInput}
\alias{wallet_connectInput}
\title{\if{html}{\out{
Wallet_Connect
}}}
\usage{
wallet_connectInput(inputId, default = "")
}
\description{
\if{html}{\out{
Integrates with Dynamic Wallet Connect widget
}}
}
