% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet_connect.R
\name{updateWallet_connectInput}
\alias{updateWallet_connectInput}
\title{\if{html}{\out{
Wallet_Connect
}}}
\usage{
updateWallet_connectInput(session, inputId, value, configuration = NULL)
}
\description{
\if{html}{\out{
Integrates with Dynamic Wallet Connect widget
}}
}
